# apps/api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/projects/{projectId}/devices/stats": {
      "get": {
        "operationId": "getStats",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
...
    "/api/v1/projects/{projectId}/notifications/send": {
      "post": {
        "description": "\n      Sends a push notification to specified devices or user segments.\n      Supports multi-platform delivery (iOS, Android, Web) with platform-specific customization.\n      Rate limited to 100 requests per minute.\n    ",
        "operationId": "send",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "example": "64f1a2b3c4d5e6f7a8b9c0d1",
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Notification sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": "64f1a2b3c4d5e6f7a8b9c0d2",
                    "projectId": "64f1a2b3c4d5e6f7a8b9c0d1",
                    "status": "sent",
                    "type": "push",
                    "title": "Welcome!",
                    "body": "Thanks for installing our app!",
                    "recipients": {
                      "total": 150,
                      "successful": 148,
                      "failed": 2
                    },
                    "platforms": {
                      "fcm": {
                        "sent": 120,
                        "delivered": 118
                      },
                      "apns": {
                        "sent": 28,
                        "delivered": 28
                      },
                      "webPush": {
                        "sent": 2,
                        "delivered": 2
                      }
                    },
                    "scheduledAt": null,
                    "sentAt": "2023-12-07T16:30:00.000Z",
                    "createdAt": "2023-12-07T16:30:00.000Z"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid notification data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "headers": {
              "X-RateLimit-Limit": {
                "description": "Request limit",
                "schema": {
                  "type": "string"
                }
              },
              "X-RateLimit-Remaining": {
                "description": "Remaining requests",
                "schema": {
                  "type": "string"
                }
              },
              "X-RateLimit-Reset": {
                "description": "Reset time",
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Send push notification",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/projects/{projectId}/notifications": {
      "get": {
        "description": "Retrieves a paginated list of notifications with optional filtering by status and type.",
        "operationId": "Notifications_List",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "description": "Project ID",
            "schema": {
              "example": "64f1a2b3c4d5e6f7a8b9c0d1",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by notification status",
            "schema": {
              "enum": [
                "pending",
                "processing",
                "sent",
                "delivered",
                "failed",
                "cancelled"
              ],
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by notification type",
            "schema": {
              "enum": [
                "instant",
                "scheduled",
                "recurring"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of notifications to return (max 100)",
            "schema": {
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Number of notifications to skip (offset)",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "example": "createdAt",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "items": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Notification"
                      }
                    },
                    "total": {
                      "type": "number",
                      "example": 1250
                    },
                    "limit": {
                      "type": "number",
                      "example": 20
                    },
                    "offset": {
                      "type": "number",
                      "example": 0
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "List notifications",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/projects/{projectId}/notifications/stats": {
      "get": {
        "operationId": "getStats",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "days",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/projects/{projectId}/notifications/{id}": {
      "get": {
        "operationId": "Notifications_GetById",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Notification not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Get notification by id",
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/v1/projects/{projectId}/notifications/{id}/cancel": {
      "patch": {
        "operationId": "Notifications_Cancel",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "description": "API Key for authentication",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "projectId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Notification not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Cancel a scheduled/recurring notification",
        "tags": [
          "Notifications"
        ]
      }
    },
...
  },
  "info": {
    "title": "Push Notification Service API",
    "description": "\n      Enterprise-grade Push Notification Service supporting iOS, Android, and Web platforms.\n      \n      ## Features\n      - 🚀 Multi-platform push notifications (FCM, APNs, Web Push)\n      - 📱 Advanced device management and segmentation\n      - 📊 Real-time analytics and delivery tracking\n      - 🏷️ Template-based notifications with variable substitution\n      - ⚡ High-performance queue system with BullMQ\n      - 🔒 Enterprise security with API key authentication\n      - 🛡️ Rate limiting for API protection\n      - 📈 99.9% uptime reliability target\n      \n      ## Authentication\n      All endpoints require an API key in the `X-API-Key` header.\n      \n      ## Rate Limiting\n      API endpoints are rate-limited to ensure fair usage:\n      - **High Frequency**: 100 requests/minute (notifications, events)\n      - **Medium Frequency**: 300 requests/minute (device management)\n      - **Low Frequency**: 1000 requests/hour (analytics, admin)\n      \n      Rate limit headers are included in all responses:\n      - `X-RateLimit-Limit`: Maximum requests allowed\n      - `X-RateLimit-Remaining`: Requests remaining in current window\n      - `X-RateLimit-Reset`: Unix timestamp when the rate limit resets\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "Push Notification Service Team",
      "url": "https://github.com/your-repo/push-notification-service",
...
  "servers": [
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API Key for project authentication. Get your API key from the project dashboard."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "object",
            "example": "Validation failed"
          },
          "timestamp": {
            "type": "string",
            "example": "2025-01-01T10:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/projects/123/notifications/send"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "timestamp",
          "path"
        ]
      },
...
      "RecurringConfigDto": {
        "type": "object",
        "properties": {
          "pattern": {
            "type": "string",
            "description": "Cron pattern for recurring notifications",
            "example": "0 9 * * MON-FRI"
          },
          "timezone": {
            "type": "string",
            "description": "Timezone for scheduling",
            "example": "America/New_York"
          },
          "endDate": {
            "type": "string",
            "description": "End date for recurring notifications",
            "example": "2024-12-31T23:59:59.000Z"
          }
        },
        "required": [
          "pattern",
          "timezone"
        ]
      },
      "SendNotificationDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Notification title",
            "example": "Welcome to our app!",
            "minLength": 1,
            "maxLength": 100
          },
          "body": {
            "type": "string",
            "description": "Notification body/message",
            "example": "Thanks for installing our app. Get started with these features!",
            "minLength": 1,
            "maxLength": 1000
          },
          "imageUrl": {
            "type": "string",
            "description": "Image URL for rich notifications",
            "example": "https://example.com/images/welcome.jpg"
          },
          "actionUrl": {
            "type": "string",
            "description": "Action URL when notification is clicked",
            "example": "https://example.com/welcome"
          },
          "data": {
            "type": "object",
            "description": "Custom data payload for the notification",
            "example": {
              "userId": "12345",
              "campaignId": "welcome_series_1",
              "customAction": "open_feature"
            }
          },
          "type": {
            "type": "string",
            "description": "Notification type",
            "enum": [
              "instant",
              "scheduled",
              "recurring"
            ],
            "example": "instant"
          },
          "targetDevices": {
            "description": "Array of specific device IDs to target",
            "example": [
              "device_123",
              "device_456"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targetTags": {
            "description": "Array of tags to target devices",
            "example": [
              "premium",
              "beta-tester",
              "ios"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targetTopics": {
            "description": "Array of topics to target",
            "example": [
              "news",
              "promotions"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "targetQuery": {
            "type": "object",
            "description": "Advanced query for targeting specific user segments",
            "example": {
              "platform": "android",
              "appVersion": {
                "$gte": "2.0.0"
              },
              "lastSeen": {
                "$gte": "2023-12-01"
              }
            }
          },
          "scheduledFor": {
            "type": "string",
            "description": "Schedule notification for future delivery",
            "example": "2023-12-08T09:00:00.000Z"
          },
          "expiresAt": {
            "type": "string",
            "description": "Expiration time for the notification",
            "example": "2023-12-08T18:00:00.000Z"
          },
          "recurring": {
            "description": "Configuration for recurring notifications",
            "allOf": [
              {
                "$ref": "#/components/schemas/RecurringConfigDto"
              }
            ]
          },
          "templateId": {
            "type": "string",
            "description": "Template ID for template-based notifications",
            "example": "64f1a2b3c4d5e6f7a8b9c0d4"
          },
          "templateVariables": {
            "type": "object",
            "description": "Variables for template substitution",
            "example": {
              "userName": "John Doe",
              "productName": "Premium Plan",
              "discountAmount": "20%"
            }
          }
        },
        "required": [
          "title",
          "body"
        ]
      },
      "Notification": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}